chore: move code around, make test fixtures import from next-mdx-remote
