+++ __tests__/fixtures/basic/pages/index.jsx
@@ -2,8 +2,8 @@ import fs from 'fs'
 import path from 'path'
 import { createContext, useEffect, useState } from 'react'
 import dynamic from 'next/dynamic'
-import { serialize } from '../../../../serialize'
-import { MDXRemote } from '../../../../'
+import { serialize } from 'next-mdx-remote/serialize'
+import { MDXRemote } from 'next-mdx-remote'
 import Test from '../components/test'
 import { paragraphCustomAlerts } from '@hashicorp/remark-plugins'
