+++ __tests__/fixtures/rsc/components/dynamic.jsx
@@ -1,0 +1,3 @@
+export default function Dynamic() {
+  return <div>I am a dynamic component.</div>
+}
