+++ __tests__/fixtures/rsc/components/test.jsx
@@ -1,0 +1,12 @@
+'use client'
+import { useState } from 'react'
+
+export default function Test({ name }) {
+  const [count, setCount] = useState(0)
+  return (
+    <>
+      <p>hello {name}</p>
+      <button onClick={() => setCount(count + 1)}>Count: {count}</button>
+    </>
+  )
+}
