+++ __tests__/fixtures/rsc/mdx/test.mdx
@@ -1,0 +1,32 @@
+---
+title: 'foo'
+name: jeff
+---
+
+import Bar from 'bar'
+
+## Headline
+
+<Test name={frontmatter.name} />
+
+<ContextConsumer />
+
+Some **markdown** content
+
+<Dynamic />
+
+```shell-session
+curl localhost
+```
+
+This is more text.
+
+"Authorize \<GITHUB_USER\>"
+
+(support for version \<230)
+
+### Some version \<= 1.3.x
+
+#### metric.name.\<operation>.\<mount>
+
+< 8ms
