+++ __tests__/fixtures/rsc/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "dependencies": {
+    "next-mdx-remote": "*"
+  }
+}
