+++ package.json
@@ -1,118 +1,7 @@
 {
-  "name": "next-mdx-remote",
-  "description": "utilities for loading mdx from any remote source as data, rather than as a local import",
-  "version": "4.2.1",
-  "author": "Jeff Escalante",
-  "bugs": {
-    "url": "https://github.com/hashicorp/next-mdx-remote/issues"
-  },
-  "exports": {
-    ".": "./index.js",
-    "./serialize": "./serialize.js",
-    "./rsc": "./rsc.js"
-  },
-  "dependencies": {
-    "@mdx-js/mdx": "^2.2.1",
-    "@mdx-js/react": "^2.2.1",
-    "vfile": "^5.3.0",
-    "vfile-matter": "^3.0.1"
-  },
-  "devDependencies": {
-    "@babel/code-frame": "^7.16.0",
-    "@hashicorp/remark-plugins": "^3.2.1",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "@rollup/plugin-typescript": "^8.3.0",
-    "@types/babel__code-frame": "^7.0.3",
-    "@types/cross-spawn": "^6.0.2",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.6",
-    "@types/puppeteer": "^5.4.4",
-    "@types/react": "^17.0.33",
-    "@types/rmfr": "^2.0.1",
-    "@types/serve-handler": "^6.1.1",
-    "cheerio": "^1.0.0-rc.10",
-    "cross-spawn": "^7.0.3",
-    "esbuild-jest": "^0.5.0",
-    "husky": "^4.3.6",
-    "jest": "^27.3.1",
-    "next": "^13.1.2",
-    "prettier": "^2.4.1",
-    "pretty-quick": "^3.1.1",
-    "puppeteer": "^10.4.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "release": "^6.3.0",
-    "rmfr": "^2.0.0",
-    "rollup": "^2.59.0",
-    "serve-handler": "^6.1.3",
-    "tslib": "^2.3.1",
-    "typescript": "^4.4.4",
-    "unist-util-remove": "^3.1.0"
-  },
-  "homepage": "https://github.com/hashicorp/next-mdx-remote#readme",
-  "husky": {
-    "hooks": {
-      "pre-commit": "pretty-quick --staged"
-    }
-  },
-  "jest": {
-    "testPathIgnorePatterns": [
-      "fixtures"
-    ],
-    "transformIgnorePatterns": [
-      "/node_modules/(?!(@mdx-js|unist-|unified|bail|is-plain-obj|trough|vfile|remark-|micromark-|micromark|estree-util-|parse-entities|character-entities|mdast-util-|character-reference-invalid|is-|stringify-entities|periscopic|hast-util-|comma-separated-tokens|property-information|space-separated-tokens|zwitch|ccount|decode-named-character-reference|trim-lines))"
-    ],
-    "transform": {
-      "\\.m?[jt]sx?$": [
-        "esbuild-jest",
-        {
-          "sourcemap": true,
-          "loaders": {
-            ".mjs": "jsx"
-          }
-        }
-      ]
-    },
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "mjs"
-    ],
-    "verbose": true
-  },
-  "keywords": [
-    "markdown",
-    "mdx",
-    "next",
-    "next.js"
-  ],
-  "license": "MPL-2.0",
-  "peerDependencies": {
-    "react": ">=16.x <=18.x",
-    "react-dom": ">=16.x <=18.x"
-  },
-  "engines": {
-    "node": ">=14",
-    "npm": ">=7"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/hashicorp/next-mdx-remote.git"
-  },
-  "scripts": {
-    "build": "rm -rf ./dist && rollup -c rollup.config.js",
-    "postbuild": "cp ./src/jsx-runtime.cjs dist",
-    "lint": "prettier --write \"./**/*.{js,jsx,json}\"",
-    "release:major": "release major && npm publish",
-    "release:minor": "release minor && npm publish",
-    "release:patch": "release patch && npm publish",
-    "prepublishOnly": "npm run build",
-    "test": "jest"
-  },
-  "types": "index.d.ts",
-  "type": "module"
+  "name": "next-mdx-remote-root",
+  "private": true,
+  "workspaces": [
+    "./packages/*"
+  ]
 }
