+++ packages/next-mdx-remote/index.d.ts
