+++ packages/next-mdx-remote/serialize.d.ts
