+++ packages/next-mdx-remote/serialize.js
