+++ packages/next-mdx-remote/src/jsx-runtime.cjs
