+++ packages/next-mdx-remote/src/serialize.ts
