+++ packages/next-mdx-remote/tsconfig.json
